packages/bridge/requirements.txt
paho-mqtt
asyncio
can
cantools
python-can == 4.0.0